updated for version 7.3.744 Problem:    64 bit compiler warning. Solution:   Add type cast. (Mike Williams)

+++ src/ex_cmds.c
@@ -6460,7 +6460,7 @@ helptags_one(dir, ext, tagfname, add_help_tags)
     /*
      * Find all *.txt files.
      */
-    dirlen = STRLEN(dir);
+    dirlen = (int)STRLEN(dir);
     STRCPY(NameBuff, dir);
     STRCAT(NameBuff, "/**/*");
     STRCAT(NameBuff, ext);
